+++ packages/generator/src/edmx-to-vdm/v4/operation.spec.ts
@@ -4,11 +4,15 @@ import {
   EdmxComplexType,
   EdmxEntitySet,
   EdmxEntityTypeV4,
-  EdmxOperation
+  EdmxOperation,
+  EdmxOperationImport
 } from '../../edmx-parser/v4';
 import { ServiceMetadata } from '../../edmx-parser';
 import { generateEntitiesV4 } from './entity';
-import { generateOperations } from './operation';
+import {
+  filterAndTransformOperations,
+  generateUnboundOperations
+} from './operation';
 import { generateComplexTypesV4 } from './complex-type';
 import { createEntityType, getComplexType, getFormatter } from './entity.spec';
 
@@ -17,11 +21,126 @@ describe('action-import', () => {
     jest.clearAllMocks();
   });
 
+  it('considers the isBound filter removing unbound', () => {
+    const imports: EdmxOperationImport[] = [{ operationName: 'op1' } as any];
+    const operations: EdmxOperation[] = [
+      { Name: 'op1', IsBound: 'false' }
+    ] as any;
+    const joined = filterAndTransformOperations(imports, operations, true);
+    expect(joined).toEqual([]);
+  });
+
+  it('considers only operations where a import matches', () => {
+    const imports: EdmxOperationImport[] = [{ operationName: 'op1' } as any];
+    const operations: EdmxOperation[] = [
+      { Name: 'op1', IsBound: 'false' },
+      { Name: 'op2', IsBound: 'false' }
+    ] as any;
+    const joined = filterAndTransformOperations(imports, operations, false);
+    expect(joined).toEqual([{ operationName: 'op1', IsBound: false }]);
+  });
+
+  it('considers only operations where a import with namespace matches', () => {
+    const imports: EdmxOperationImport[] = [{ operationName: 'ns.op1' } as any];
+    const operations: EdmxOperation[] = [
+      { Name: 'op1', IsBound: 'false' },
+      { Name: 'op2', IsBound: 'false' }
+    ] as any;
+    const joined = filterAndTransformOperations(imports, operations, false);
+    expect(joined).toEqual([{ operationName: 'ns.op1', IsBound: false }]);
+  });
+
+  it('removes bound operations with no parameter', () => {
+    const imports: EdmxOperationImport[] = [
+      { operationName: 'op1' },
+      { operationName: 'op2' }
+    ] as any;
+    const operations: EdmxOperation[] = [
+      { Name: 'op1', Parameter: [], IsBound: 'true' },
+      { Name: 'op2', Parameter: [{ Type: 'ns.op2' }], IsBound: 'true' }
+    ] as any;
+    const joined = filterAndTransformOperations(imports, operations, true);
+    expect(joined).toEqual([
+      {
+        operationName: 'op2',
+        entitySetName: 'op2',
+        IsBound: true,
+        Parameter: []
+      }
+    ]);
+  });
+
+  it('considers bound operations', () => {
+    const imports: EdmxOperationImport[] = [
+      { operationName: 'op1' },
+      { operationName: 'op2' }
+    ] as any;
+    const operations: EdmxOperation[] = [
+      { Name: 'op1', Parameter: [{ Type: 'ns.Entity' }], IsBound: 'true' },
+      { Name: 'op2', IsBound: 'true', Parameter: [] }
+    ] as any;
+    const joined = filterAndTransformOperations(imports, operations, true);
+    expect(joined).toEqual([
+      {
+        IsBound: true,
+        operationName: 'op1',
+        Parameter: [],
+        entitySetName: 'Entity'
+      }
+    ]);
+  });
+
+  it('handles bound operations with multiple dots in the name', () => {
+    const imports: EdmxOperationImport[] = [
+      { operationName: 'op1' },
+      { operationName: 'op2' }
+    ] as any;
+    const operations: EdmxOperation[] = [
+      {
+        Name: 'op1',
+        Parameter: [{ Type: 'ns.foo.bar.baz.Entity' }],
+        IsBound: 'true'
+      },
+      { Name: 'op2', IsBound: 'true', Parameter: [] }
+    ] as any;
+    const joined = filterAndTransformOperations(imports, operations, true);
+    expect(joined).toEqual([
+      {
+        IsBound: true,
+        operationName: 'op1',
+        Parameter: [],
+        entitySetName: 'Entity'
+      }
+    ]);
+  });
+
+  it('considers isBound filter removing bound operations', () => {
+    const imports: EdmxOperationImport[] = [{ operationName: 'op1' }] as any;
+    const operations: EdmxOperation[] = [
+      { Name: 'op1', Parameter: [{ Type: 'ns.Entity' }], IsBound: 'true' }
+    ] as any;
+    const joined = filterAndTransformOperations(imports, operations, false);
+    expect(joined).toEqual([]);
+  });
+
+  it('removes first parameter for bound operations', () => {
+    const imports: EdmxOperationImport[] = [{ operationName: 'op1' }] as any;
+    const operations: EdmxOperation[] = [
+      {
+        Name: 'op1',
+        Parameter: [{ Type: 'ns.Entity' }, { Type: 'para1' }],
+        IsBound: 'true'
+      }
+    ] as any;
+    const joined = filterAndTransformOperations(imports, operations, true);
+    expect(joined[0].Parameter).toEqual([{ Type: 'para1' }]);
+  });
+
   it('transforms action imports', () => {
     const formatter = getFormatter();
     const service = createServiceWithActions();
     const entities = generateEntitiesV4(service, [], [], formatter);
-    const actionImport = generateOperations(
+    const actionImport = generateUnboundOperations(
       service,
       'myServiceWithActions',
       'action',
@@ -104,7 +223,7 @@ describe('action-import', () => {
     );
     expect(entities.length).toBe(2);
 
-    const actionImports = generateOperations(
+    const actionImports = generateUnboundOperations(
       service,
       'myTestServiceName',
       'action',
@@ -122,7 +241,7 @@ describe('action-import', () => {
     const formatter = getFormatter();
     const service =
       createServiceMetadataWithActionImportLinksToUndefinedAction();
-    generateOperations(
+    generateUnboundOperations(
       service,
       'myTestServiceName',
       'action',
@@ -144,7 +263,7 @@ describe('action-import', () => {
     const formatter = getFormatter();
     const service = createServiceWithActions();
     const entities = generateEntitiesV4(service, [], [], formatter);
-    generateOperations(
+    generateUnboundOperations(
       service,
       'myTestServiceName',
       'action',
@@ -173,7 +292,7 @@ function createAction(
           Nullable: 'false'
         }
       : undefined,
-    IsBound: false,
+    IsBound: 'false',
     Namespace: namespace
   };
 }