+++ packages/generator/src/generator.spec.ts
@@ -181,7 +181,7 @@ describe('generator', () => {
 
       expect(testEntityFile).toBeDefined();
       expect(testEntityFile!.getClasses().length).toBe(1);
-      expect(testEntityFile!.getInterfaces().length).toBe(1);
+      expect(testEntityFile!.getInterfaces().length).toBe(4);
       const imports = testEntityFile!
         .getImportStringLiterals()
         .map(stringLiteral => stringLiteral.getLiteralValue());