+++ packages/generator/src/operations/import.ts
@@ -90,8 +90,35 @@ export function operationImportDeclarations(
   type: 'function' | 'action',
   operations: VdmOperation[] = []
 ): ImportDeclarationStructure[] {
+  if (!operations.length) {
+    return [];
+  }
+
   const parameters = operations.flatMap(({ parameters: params }) => params);
   const returnTypes = operations.map(({ returnType }) => returnType);
+
+  const includesBound = !!operations.filter(operation => operation.isBound)
+    .length;
+  const includesUnbound = !!operations.filter(operation => !operation.isBound)
+    .length;
+
+  const hasOperationWithParameters = !!operations.filter(
+    operation => operation.parameters.length > 0 && operation.type === type
+  ).length;
+  if (includesUnbound && includesBound) {
+    throw new Error(
+      'Bound and unbound operations found in generation - this should not happen.'
+    );
+  }
+  const serviceImport: ImportDeclarationStructure[] = includesBound
+    ? []
+    : [
+        {
+          kind: StructureKind.ImportDeclaration,
+          namedImports: [voca.decapitalize(className)],
+          moduleSpecifier: './service'
+        }
+      ];
 
   return [
     ...externalImportDeclarations(parameters),
@@ -104,16 +131,19 @@ export function operationImportDeclarations(
         'DefaultDeSerializers',
         'defaultDeSerializers',
         ...propertyTypeImportNames(parameters),
-        `${voca.capitalize(type)}ImportParameter`,
-        `${voca.capitalize(type)}ImportRequestBuilder`
+        ...(hasOperationWithParameters
+          ? [`${voca.capitalize(type)}ImportParameter`]
+          : []),
+        ...(includesUnbound
+          ? [`${voca.capitalize(type)}ImportRequestBuilder`]
+          : []),
+        ...(includesBound
+          ? [`Bound${voca.capitalize(type)}ImportRequestBuilder`]
+          : [])
       ],
       oDataVersion
     ),
-    {
-      kind: StructureKind.ImportDeclaration,
-      namedImports: [voca.decapitalize(className)],
-      moduleSpecifier: './service'
-    },
+    ...serviceImport,
     ...returnTypeImports(returnTypes)
   ];
 }