+++ packages/generator/src/operations/operation.spec.ts
@@ -8,6 +8,36 @@ import {
 import { operationFunction } from './operation';
 
 describe('function', () => {
+  it('creates statement for bound operations', () => {
+    const actual = operationFunction(
+      {
+        ...orderBreakfast,
+        isBound: true,
+        entityClassName: 'BreakfastEntity'
+      },
+      { ...foodService, oDataVersion: 'v4' }
+    );
+    expect(actual).toEqual({
+      kind: StructureKind.Function,
+      name: 'orderBreakfast',
+      parameters: [
+        { name: 'parameters', type: 'Params<T>' },
+        {
+          initializer: undefined,
+          name: 'deSerializers?',
+          type: 'T'
+        }
+      ],
+      returnType:
+        'BoundFunctionImportRequestBuilder<BreakfastEntity<T>, T, Params<T>, string>',
+      docs: [
+        `order a breakfast ${unixEOL}@param parameters - Object containing all parameters for the function.${unixEOL}@returns A request builder that allows to overwrite some of the values and execute the resulting request.`
+      ],
+      isExported: true,
+      statements: `const params = {${unixEOL}withHoneyToast: new FunctionImportParameter('WithHoneyToast', 'Edm.Boolean', parameters.withHoneyToast)${unixEOL}};${unixEOL}${unixEOL}return new BoundFunctionImportRequestBuilder(this._entityApi, this, 'OrderBreakfast', (data) => transformReturnValueForEdmType(data, (val) => edmToTs(val, 'Edm.String', deSerializers)), params, deSerializers || defaultDeSerializers);`
+    });
+  });
+
   it('functionImportFunction', () => {
     const actual = operationFunction(orderBreakfast, foodService);
     expect(actual).toEqual({