+++ packages/odata-common/src/request-builder/create-request-builder-base.ts
@@ -1,6 +1,7 @@
 import { ErrorWithCause } from '@sap-cloud-sdk/util';
 import { DestinationOrFetchOptions } from '@sap-cloud-sdk/connectivity';
 import { HttpResponse } from '@sap-cloud-sdk/http-client';
+import { v4 as uuid } from 'uuid';
 import type { EntitySerializer } from '../entity-serializer';
 import type { ODataUri } from '../uri-conversion';
 import type { EntityBase, EntityIdentifiable } from '../entity-base';
@@ -10,6 +11,10 @@ import { ODataCreateRequestConfig } from '../request/odata-create-request-config
 import { Link } from '../selectable';
 import { DeSerializers } from '../de-serializers/de-serializers';
 import { EntityApi } from '../entity-api';
+import {
+  BatchReference,
+  WithBatchReference
+} from '../request/odata-request-traits';
 import { MethodRequestBuilder } from './request-builder-base';
 
 /**
@@ -23,9 +28,10 @@ export abstract class CreateRequestBuilderBase<
   extends MethodRequestBuilder<
     ODataCreateRequestConfig<EntityT, DeSerializersT>
   >
-  implements EntityIdentifiable<EntityT, DeSerializersT>
+  implements EntityIdentifiable<EntityT, DeSerializersT>, WithBatchReference
 {
   readonly _deSerializers: DeSerializersT;
+  private _batchReference: BatchReference;
 
   /**
    * Creates an instance of CreateRequestBuilder.
@@ -53,6 +59,25 @@ export abstract class CreateRequestBuilderBase<
 
   get entity(): EntityT {
     return this._entity;
+  }
+
+  /**
+   * Gets identifier for the batch request.
+   * @returns Batch request identifier.
+   */
+  getBatchReference(): BatchReference {
+    if (!this._batchReference) {
+      this.setBatchId(uuid());
+    }
+    return this._batchReference;
+  }
+
+  /**
+   * Sets user-defined identifier for the batch reference.
+   * @param id - User-defined batch reuest identifier.
+   */
+  setBatchId(id: string): void {
+    this._batchReference = { id };
   }
 
   /**