+++ test-packages/test-services-e2e/v4/test-service/TestEntity.d.ts
@@ -7,9 +7,12 @@ import {
   Entity,
   DefaultDeSerializers,
   DeSerializers,
-  DeserializedType
+  DeserializedType,
+  BoundActionImportRequestBuilder,
+  BoundFunctionImportRequestBuilder
 } from '@sap-cloud-sdk/odata-v4';
 import type { TestEntityApi } from './TestEntityApi';
+import { MyComplexReturnType } from './MyComplexReturnType';
 import { TestEntityLink, TestEntityLinkType } from './TestEntityLink';
 /**
  * This class represents the entity "TestEntity" of service "TestService".
@@ -97,6 +100,62 @@ export declare class TestEntity<T extends DeSerializers = DefaultDeSerializers>
    */
   toMultiLink: TestEntityLink<T>[];
   constructor(_entityApi: TestEntityApi<T>);
+  /**
+   * Bound Function Without Arguments.
+   * @param parameters - Object containing all parameters for the function.
+   * @returns A request builder that allows to overwrite some of the values and execute the resulting request.
+   */
+  boundFunctionWithoutArguments(
+    parameters: BoundFunctionWithoutArgumentsParameters<T>,
+    deSerializers?: T
+  ): BoundFunctionImportRequestBuilder<
+    TestEntity<T>,
+    T,
+    BoundFunctionWithoutArgumentsParameters<T>,
+    string | null
+  >;
+  /**
+   * Bound Function Without Arguments Complex Return Type.
+   * @param parameters - Object containing all parameters for the function.
+   * @returns A request builder that allows to overwrite some of the values and execute the resulting request.
+   */
+  boundFunctionWithoutArgumentsComplexReturnType(
+    parameters: BoundFunctionWithoutArgumentsComplexReturnTypeParameters<T>,
+    deSerializers?: T
+  ): BoundFunctionImportRequestBuilder<
+    TestEntity<T>,
+    T,
+    BoundFunctionWithoutArgumentsComplexReturnTypeParameters<T>,
+    MyComplexReturnType | null
+  >;
+  /**
+   * Bound Function With Arguments.
+   * @param parameters - Object containing all parameters for the function.
+   * @returns A request builder that allows to overwrite some of the values and execute the resulting request.
+   */
+  boundFunctionWithArguments(
+    parameters: BoundFunctionWithArgumentsParameters<T>,
+    deSerializers?: T
+  ): BoundFunctionImportRequestBuilder<
+    TestEntity<T>,
+    T,
+    BoundFunctionWithArgumentsParameters<T>,
+    string | null
+  >;
+  /**
+   * Bound Action Without Arguments.
+   * @param parameters - Object containing all parameters for the action.
+   * @returns A request builder that allows to overwrite some of the values and execute the resulting request.
+   */
+  boundActionWithoutArguments(
+    parameters: BoundActionWithoutArgumentsParameters<T>,
+    deSerializers?: T
+  ): BoundActionImportRequestBuilder<
+    TestEntity<T>,
+    T,
+    BoundActionWithoutArgumentsParameters<T>,
+    string | null
+  >;
 }
 export interface TestEntityType<
   T extends DeSerializers = DefaultDeSerializers
@@ -120,4 +179,37 @@ export interface TestEntityType<
   > | null;
   toMultiLink: TestEntityLinkType<T>[];
 }
+/**
+ * Type of the parameters to be passed to {@link boundFunctionWithoutArguments}.
+ */
+export interface BoundFunctionWithoutArgumentsParameters<
+  DeSerializersT extends DeSerializers
+> {}
+/**
+ * Type of the parameters to be passed to {@link boundFunctionWithoutArgumentsComplexReturnType}.
+ */
+export interface BoundFunctionWithoutArgumentsComplexReturnTypeParameters<
+  DeSerializersT extends DeSerializers
+> {}
+/**
+ * Type of the parameters to be passed to {@link boundFunctionWithArguments}.
+ */
+export interface BoundFunctionWithArgumentsParameters<
+  DeSerializersT extends DeSerializers
+> {
+  /**
+   * Param 1.
+   */
+  param1?: string | null;
+  /**
+   * Param 2.
+   */
+  param2?: string | null;
+}
+/**
+ * Type of the parameters to be passed to {@link boundActionWithoutArguments}.
+ */
+export interface BoundActionWithoutArgumentsParameters<
+  DeSerializersT extends DeSerializers
+> {}
 //# sourceMappingURL=TestEntity.d.ts.map