+++ test-packages/test-services-e2e/v4/test-service/TestEntity.js
@@ -7,6 +7,7 @@ exports.TestEntity = void 0;
  * This is a generated file powered by the SAP Cloud SDK for JavaScript.
  */
 const odata_v4_1 = require("@sap-cloud-sdk/odata-v4");
+const MyComplexReturnType_1 = require("./MyComplexReturnType");
 /**
  * This class represents the entity "TestEntity" of service "TestService".
  */
@@ -14,6 +15,45 @@ class TestEntity extends odata_v4_1.Entity {
     constructor(_entityApi) {
         super(_entityApi);
         this._entityApi = _entityApi;
+    }
+    /**
+     * Bound Function Without Arguments.
+     * @param parameters - Object containing all parameters for the function.
+     * @returns A request builder that allows to overwrite some of the values and execute the resulting request.
+     */
+    boundFunctionWithoutArguments(parameters, deSerializers) {
+        const params = {};
+        return new odata_v4_1.BoundFunctionImportRequestBuilder(this._entityApi, this, 'boundFunctionWithoutArguments', data => (0, odata_v4_1.transformReturnValueForEdmType)(data, val => (0, odata_v4_1.edmToTs)(val.value, 'Edm.String', deSerializers || odata_v4_1.defaultDeSerializers)), params, deSerializers || odata_v4_1.defaultDeSerializers);
+    }
+    /**
+     * Bound Function Without Arguments Complex Return Type.
+     * @param parameters - Object containing all parameters for the function.
+     * @returns A request builder that allows to overwrite some of the values and execute the resulting request.
+     */
+    boundFunctionWithoutArgumentsComplexReturnType(parameters, deSerializers) {
+        const params = {};
+        return new odata_v4_1.BoundFunctionImportRequestBuilder(this._entityApi, this, 'boundFunctionWithoutArgumentsComplexReturnType', data => (0, odata_v4_1.transformReturnValueForComplexType)(data, data => (0, odata_v4_1.entityDeserializer)(deSerializers || odata_v4_1.defaultDeSerializers).deserializeComplexType(data, MyComplexReturnType_1.MyComplexReturnType)), params, deSerializers || odata_v4_1.defaultDeSerializers);
+    }
+    /**
+     * Bound Function With Arguments.
+     * @param parameters - Object containing all parameters for the function.
+     * @returns A request builder that allows to overwrite some of the values and execute the resulting request.
+     */
+    boundFunctionWithArguments(parameters, deSerializers) {
+        const params = {
+            param1: new odata_v4_1.FunctionImportParameter('param1', 'Edm.String', parameters.param1),
+            param2: new odata_v4_1.FunctionImportParameter('param2', 'Edm.String', parameters.param2)
+        };
+        return new odata_v4_1.BoundFunctionImportRequestBuilder(this._entityApi, this, 'boundFunctionWithArguments', data => (0, odata_v4_1.transformReturnValueForEdmType)(data, val => (0, odata_v4_1.edmToTs)(val.value, 'Edm.String', deSerializers || odata_v4_1.defaultDeSerializers)), params, deSerializers || odata_v4_1.defaultDeSerializers);
+    }
+    /**
+     * Bound Action Without Arguments.
+     * @param parameters - Object containing all parameters for the action.
+     * @returns A request builder that allows to overwrite some of the values and execute the resulting request.
+     */
+    boundActionWithoutArguments(parameters, deSerializers) {
+        const params = {};
+        return new odata_v4_1.BoundActionImportRequestBuilder(this._entityApi, this, 'boundActionWithoutArguments', data => (0, odata_v4_1.transformReturnValueForEdmType)(data, val => (0, odata_v4_1.edmToTs)(val.value, 'Edm.String', deSerializers || odata_v4_1.defaultDeSerializers)), params, deSerializers || odata_v4_1.defaultDeSerializers);
     }
 }
 exports.TestEntity = TestEntity;