+++ test-packages/test-services-e2e/v4/test-service/service.d.ts
@@ -4,12 +4,14 @@
  * This is a generated file powered by the SAP Cloud SDK for JavaScript.
  */
 import { TestEntityApi } from './TestEntityApi';
+import { TestEntityWithMultipleKeysApi } from './TestEntityWithMultipleKeysApi';
 import { TestEntityLinkApi } from './TestEntityLinkApi';
 import { TestEntity50PropApi } from './TestEntity50PropApi';
 import {
   ConcatStringsParameters,
   GetAllParameters,
   GetByKeyParameters,
+  GetByKeyWithMultipleKeysParameters,
   ReturnCollectionParameters,
   ReturnIntParameters,
   ReturnSapCloudSdkParameters
@@ -101,6 +103,7 @@ declare class TestService<
   constructor(deSerializers: DeSerializersT);
   private initApi;
   get testEntityApi(): TestEntityApi<DeSerializersT>;
+  get testEntityWithMultipleKeysApi(): TestEntityWithMultipleKeysApi<DeSerializersT>;
   get testEntityLinkApi(): TestEntityLinkApi<DeSerializersT>;
   get testEntity50PropApi(): TestEntity50PropApi<DeSerializersT>;
   get functionImports(): {
@@ -124,6 +127,13 @@ declare class TestService<
       DeSerializersT,
       GetByKeyParameters<DeSerializersT>,
       import('./TestEntity').TestEntity<DefaultDeSerializers>
+    >;
+    getByKeyWithMultipleKeys: (
+      parameter: GetByKeyWithMultipleKeysParameters<DeSerializersT>
+    ) => import('@sap-cloud-sdk/odata-v4').FunctionImportRequestBuilder<
+      DeSerializersT,
+      GetByKeyWithMultipleKeysParameters<DeSerializersT>,
+      import('./TestEntityWithMultipleKeys').TestEntityWithMultipleKeys<DefaultDeSerializers>
     >;
     returnCollection: (
       parameter: ReturnCollectionParameters<DeSerializersT>